+++ manager/pkg/spec/controllers/bundle/base_objects_bundle.go
