+++ manager/pkg/spec/controllers/bundle/interface.go
