+++ manager/pkg/spec/controllers/managedclustersetbinding_controller.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Red Hat, Inc.
 // Copyright Contributors to the Open Cluster Management project
 
-package todatabase
+package controllers
 
 import (
 	"fmt"
