+++ manager/pkg/spec/specdb/gorm/spec.go
@@ -10,7 +10,7 @@ import (
 	"github.com/jackc/pgx/v4"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/spec/todatabase/bundle"
+	"github.com/stolostron/multicluster-global-hub/manager/pkg/spec/controllers/bundle"
 	"github.com/stolostron/multicluster-global-hub/pkg/database"
 )
