+++ manager/pkg/spec/syncers/interval/interval_policy.go
