+++ manager/pkg/spec/syncers/placementbindings_syncer.go
@@ -1,4 +1,4 @@
-package syncer
+package syncers
 
 import (
 	"context"
@@ -9,9 +9,9 @@ import (
 	policyv1 "open-cluster-management.io/governance-policy-propagator/api/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 
+	"github.com/stolostron/multicluster-global-hub/manager/pkg/spec/controllers/bundle"
 	"github.com/stolostron/multicluster-global-hub/manager/pkg/spec/specdb"
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/spec/todatabase/bundle"
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/spec/totransport/interval"
+	"github.com/stolostron/multicluster-global-hub/manager/pkg/spec/syncers/interval"
 	"github.com/stolostron/multicluster-global-hub/pkg/transport"
 )
